Fixed a few crashes, better support for higan
Fixed AudioManager/InputManager unsigned vs signed comparison warnings.
Fixed a FolderData sorting crash (I can't believe nobody's reported
this).
Fixed a GuiTheme crash for empty paths.
Added the %ROM_RAW% tag, for the unescaped ROM name - useful for higan
on windows.
SystemData will now add folders that end in EXTENSION as GameDatas, and
not recurse through them.  Also useful for higan.

+++ src/AudioManager.cpp
@@ -22,7 +22,7 @@ void AudioManager::mixAudio(void *unused, Uint8 *stream, int len)
 		{
 			//calculate rest length of current sample
 			Uint32 restLength = (sound->getLength() - sound->getPosition());
-			if (restLength > len) {
+			if (restLength > (Uint32)len) {
 				//if stream length is smaller than smaple lenght, clip it
 				restLength = len;
 			}
